Models: Move installation related models into their own file
