+++ rare/components/tabs/downloads/__init__.py
@@ -20,7 +20,7 @@ from rare.components.dialogs.install_dialog import InstallDialog
 from rare.components.tabs.downloads.dl_queue_widget import DlQueueWidget, DlWidget
 from rare.components.tabs.downloads.download_thread import DownloadThread
 from rare.ui.components.tabs.downloads.downloads_tab import Ui_DownloadsTab
-from rare.utils.models import InstallOptionsModel, InstallQueueItemModel
+from rare.models.install import InstallOptionsModel, InstallQueueItemModel
 from rare.utils.utils import get_size
 
 logger = getLogger("Download")
