+++ rare/models/signals.py
@@ -1,6 +1,6 @@
 from PyQt5.QtCore import QObject, pyqtSignal
 
-from rare.utils.models import InstallOptionsModel
+from .install import InstallOptionsModel
 
 
 class GlobalSignals(QObject):
